rename shoot-client to vpn-client, seed-server to vpn-server
