+++ Makefile
@@ -6,13 +6,12 @@ GARDENER_HACK_DIR    		  := $(shell go list -m -f "{{.Dir}}" github.com/gardener
 VERSION                       := $(shell cat VERSION)
 REPO_ROOT                     := $(shell dirname $(realpath $(lastword $(MAKEFILE_LIST))))
 REGISTRY                      := europe-docker.pkg.dev/gardener-project/public/gardener
-PREFIX                        := vpn
-SEED_SERVER_IMAGE_REPOSITORY  := $(REGISTRY)/$(PREFIX)-seed-server-go
-SEED_SERVER_IMAGE_TAG         := $(VERSION)
-LOCAL_SEED_SERVER_IMAGE_REPO  := localhost:5001/$(subst /,_,$(subst .,_,$(SEED_SERVER_IMAGE_REPOSITORY)))
-SHOOT_CLIENT_IMAGE_REPOSITORY := $(REGISTRY)/$(PREFIX)-shoot-client-go
-SHOOT_CLIENT_IMAGE_TAG        := $(VERSION)
-LOCAL_SHOOT_CLIENT_IMAGE_REPO := localhost:5001/$(subst /,_,$(subst .,_,$(SHOOT_CLIENT_IMAGE_REPOSITORY)))
+VPN_SERVER_IMAGE_REPOSITORY  := $(REGISTRY)/vpn-server
+VPN_SERVER_IMAGE_TAG         := $(VERSION)
+LOCAL_VPN_SERVER_IMAGE_REPO  := localhost:5001/$(subst /,_,$(subst .,_,$(VPN_SERVER_IMAGE_REPOSITORY)))
+VPN_CLIENT_IMAGE_REPOSITORY := $(REGISTRY)/vpn-client
+VPN_CLIENT_IMAGE_TAG        := $(VERSION)
+LOCAL_VPN_CLIENT_IMAGE_REPO := localhost:5001/$(subst /,_,$(subst .,_,$(VPN_CLIENT_IMAGE_REPOSITORY)))
 LD_FLAGS                      := "-w $(shell bash $(GARDENER_HACK_DIR)/get-build-ld-flags.sh k8s.io/component-base $(REPO_ROOT)/VERSION "vpn2")"
 
 IMAGE_TAG             := $(VERSION)
@@ -27,31 +26,31 @@ export PATH
 tidy:
 	@GO111MODULE=on go mod tidy
 
-.PHONY: seed-server-docker-image
-seed-server-docker-image:
-	@docker buildx build --platform=linux/$(ARCH) --build-arg DEBUG=$(DEBUG) -t $(SEED_SERVER_IMAGE_REPOSITORY):$(SEED_SERVER_IMAGE_TAG) -f Dockerfile --target seed-server --rm .
+.PHONY: vpn-server-docker-image
+vpn-server-docker-image:
+	@docker buildx build --platform=linux/$(ARCH) --build-arg DEBUG=$(DEBUG) -t $(VPN_SERVER_IMAGE_REPOSITORY):$(VPN_SERVER_IMAGE_TAG) -f Dockerfile --target vpn-server --rm .
 
-.PHONY: shoot-client-docker-image
-shoot-client-docker-image:
-	@docker buildx build --platform=linux/$(ARCH) --build-arg DEBUG=$(DEBUG) -t $(SHOOT_CLIENT_IMAGE_REPOSITORY):$(SHOOT_CLIENT_IMAGE_TAG) -f Dockerfile --target shoot-client --rm .
+.PHONY: vpn-client-docker-image
+vpn-client-docker-image:
+	@docker buildx build --platform=linux/$(ARCH) --build-arg DEBUG=$(DEBUG) -t $(VPN_CLIENT_IMAGE_REPOSITORY):$(VPN_CLIENT_IMAGE_TAG) -f Dockerfile --target vpn-client --rm .
 
-.PHONY: seed-server-to-gardener-local
-seed-server-to-gardener-local: seed-server-docker-image
-	@docker tag $(SEED_SERVER_IMAGE_REPOSITORY):$(SEED_SERVER_IMAGE_TAG) $(LOCAL_SEED_SERVER_IMAGE_REPO):$(SEED_SERVER_IMAGE_TAG)
-	@docker push $(LOCAL_SEED_SERVER_IMAGE_REPO):$(SEED_SERVER_IMAGE_TAG)
-	@echo "seed server image: $(LOCAL_SEED_SERVER_IMAGE_REPO):$(SEED_SERVER_IMAGE_TAG)"
+.PHONY: vpn-server-to-gardener-local
+vpn-server-to-gardener-local: vpn-server-docker-image
+	@docker tag $(VPN_SERVER_IMAGE_REPOSITORY):$(VPN_SERVER_IMAGE_TAG) $(LOCAL_VPN_SERVER_IMAGE_REPO):$(VPN_SERVER_IMAGE_TAG)
+	@docker push $(LOCAL_VPN_SERVER_IMAGE_REPO):$(VPN_SERVER_IMAGE_TAG)
+	@echo "VPN server image: $(LOCAL_VPN_SERVER_IMAGE_REPO):$(VPN_SERVER_IMAGE_TAG)"
 
-.PHONY: shoot-client-to-gardener-local
-shoot-client-to-gardener-local: shoot-client-docker-image
-	@docker tag $(SHOOT_CLIENT_IMAGE_REPOSITORY):$(SHOOT_CLIENT_IMAGE_TAG) $(LOCAL_SHOOT_CLIENT_IMAGE_REPO):$(SHOOT_CLIENT_IMAGE_TAG)
-	@docker push $(LOCAL_SHOOT_CLIENT_IMAGE_REPO):$(SHOOT_CLIENT_IMAGE_TAG)
-	@echo "shoot client image: $(LOCAL_SHOOT_CLIENT_IMAGE_REPO):$(SHOOT_CLIENT_IMAGE_TAG)"
+.PHONY: vpn-client-to-gardener-local
+vpn-client-to-gardener-local: vpn-client-docker-image
+	@docker tag $(VPN_CLIENT_IMAGE_REPOSITORY):$(VPN_CLIENT_IMAGE_TAG) $(LOCAL_VPN_CLIENT_IMAGE_REPO):$(VPN_CLIENT_IMAGE_TAG)
+	@docker push $(LOCAL_VPN_CLIENT_IMAGE_REPO):$(VPN_CLIENT_IMAGE_TAG)
+	@echo "VPN client image: $(LOCAL_VPN_CLIENT_IMAGE_REPO):$(VPN_CLIENT_IMAGE_TAG)"
 
 .PHONY: docker-images
-docker-images: seed-server-docker-image shoot-client-docker-image
+docker-images: vpn-server-docker-image vpn-client-docker-image
 
 .PHONY: docker-images-to-gardener-local
-docker-images-to-gardener-local: seed-server-to-gardener-local shoot-client-to-gardener-local
+docker-images-to-gardener-local: vpn-server-to-gardener-local vpn-client-to-gardener-local
 
 .PHONY: release
 release: docker-images docker-login docker-push
@@ -62,10 +61,10 @@ docker-login:
 
 .PHONY: docker-push
 docker-push:
-	@if ! docker images $(SEED_SERVER_IMAGE_REPOSITORY) | awk '{ print $$2 }' | grep -q -F $(SEED_SERVER_IMAGE_TAG); then echo "$(SEED_SERVER_IMAGE_REPOSITORY) version $(SEED_SERVER_IMAGE_TAG) is not yet built. Please run 'make seed-server-docker-image'"; false; fi
-	@if ! docker images $(SHOOT_CLIENT_IMAGE_REPOSITORY) | awk '{ print $$2 }' | grep -q -F $(SHOOT_CLIENT_IMAGE_TAG); then echo "$(SHOOT_CLIENT_IMAGE_REPOSITORY) version $(SHOOT_CLIENT_IMAGE_TAG) is not yet built. Please run 'make shoot-client-docker-image'"; false; fi
-	@gcloud docker -- push $(SEED_SERVER_IMAGE_REPOSITORY):$(SEED_SERVER_IMAGE_TAG)
-	@gcloud docker -- push $(SHOOT_CLIENT_IMAGE_REPOSITORY):$(SHOOT_CLIENT_IMAGE_TAG)
+	@if ! docker images $(VPN_SERVER_IMAGE_REPOSITORY) | awk '{ print $$2 }' | grep -q -F $(VPN_SERVER_IMAGE_TAG); then echo "$(VPN_SERVER_IMAGE_REPOSITORY) version $(VPN_SERVER_IMAGE_TAG) is not yet built. Please run 'make vpn-server-docker-image'"; false; fi
+	@if ! docker images $(VPN_CLIENT_IMAGE_REPOSITORY) | awk '{ print $$2 }' | grep -q -F $(VPN_CLIENT_IMAGE_TAG); then echo "$(VPN_CLIENT_IMAGE_REPOSITORY) version $(VPN_CLIENT_IMAGE_TAG) is not yet built. Please run 'make vpn-client-docker-image'"; false; fi
+	@gcloud docker -- push $(VPN_SERVER_IMAGE_REPOSITORY):$(VPN_SERVER_IMAGE_TAG)
+	@gcloud docker -- push $(VPN_CLIENT_IMAGE_REPOSITORY):$(VPN_CLIENT_IMAGE_TAG)
 
 .PHONY: check
 check:
@@ -77,16 +76,16 @@ test:
 	go test ./...
 
 .PHONY: build
-build: build-seed-server build-shoot-client
+build: build-vpn-server build-vpn-client
 
-.PHONY: build-seed-server
-build-seed-server:
-	@CGO_ENABLED=0 GOOS=linux GOARCH=$(ARCH) go build -o bin/seed-server  \
+.PHONY: build-vpn-server
+build-vpn-server:
+	@CGO_ENABLED=0 GOOS=linux GOARCH=$(ARCH) go build -o bin/vpn-server  \
 	    -ldflags $(LD_FLAGS)\
-	    ./cmd/seed_server/main.go
+	    ./cmd/vpn_server/main.go
 
-.PHONY: build-shoot-client
-build-shoot-client:
-	@CGO_ENABLED=0 GOOS=linux GOARCH=$(ARCH) go build -o bin/shoot-client  \
+.PHONY: build-vpn-client
+build-vpn-client:
+	@CGO_ENABLED=0 GOOS=linux GOARCH=$(ARCH) go build -o bin/vpn-client  \
 	    -ldflags $(LD_FLAGS)\
-	    ./cmd/shoot_client/main.go
+	    ./cmd/vpn_client/main.go
