+++ cmd/vpn_client/app/pathcontroller/ping.go
