+++ pkg/openvpn/assets/server-config.template
@@ -54,8 +54,8 @@ dev {{ .Device }}
 {{/* Add firewall rules to block all traffic originating from the shoot cluster.
      The scripts are run after the tun device has been created (up) or removed (down). */ -}}
 script-security 2
-up "/bin/seed-server firewall --mode up --device {{ .Device }}"
-down "/bin/seed-server firewall --mode down --device {{ .Device }}"
+up "/bin/vpn-server firewall --mode up --device {{ .Device }}"
+down "/bin/vpn-server firewall --mode down --device {{ .Device }}"
 
 {{ if not (eq .StatusPath "") -}}
 status {{ .StatusPath }} 15
