+++ pkg/openvpn/config_server_test.go
@@ -76,8 +76,8 @@ route 10.0.1.0 255.255.255.0
 
 			Expect(content).To(ContainSubstring(`
 script-security 2
-up "/bin/seed-server firewall --mode up --device tun0"
-down "/bin/seed-server firewall --mode down --device tun0"`))
+up "/bin/vpn-server firewall --mode up --device tun0"
+down "/bin/vpn-server firewall --mode down --device tun0"`))
 		})
 
 		It("should generate correct openvpn.config for IPv4 default values with HA", func() {
@@ -108,8 +108,8 @@ dev tap0
 
 			Expect(content).To(ContainSubstring(`
 script-security 2
-up "/bin/seed-server firewall --mode up --device tap0"
-down "/bin/seed-server firewall --mode down --device tap0"`))
+up "/bin/vpn-server firewall --mode up --device tap0"
+down "/bin/vpn-server firewall --mode down --device tap0"`))
 
 			Expect(content).To(ContainSubstring(`
 status /srv/status/openvpn.status 15
@@ -135,8 +135,8 @@ dev tun0
 `))
 			Expect(content).To(ContainSubstring(`
 script-security 2
-up "/bin/seed-server firewall --mode up --device tun0"
-down "/bin/seed-server firewall --mode down --device tun0"`))
+up "/bin/vpn-server firewall --mode up --device tun0"
+down "/bin/vpn-server firewall --mode down --device tun0"`))
 		})
 	})
